Added Create endpoint and its logic

+++ src/main/java/com/example/demo/DemoController.java
@@ -1,9 +1,11 @@
 package com.example.demo;
 
 import com.example.demo.dto.DemoResponse;
+import com.example.demo.dto.Item.CreateItem;
 import com.example.demo.helper.DebugHelp;
 import com.example.demo.model.Item;
 import com.example.demo.services.DemoService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,11 +17,11 @@ import java.util.Optional;
 @SuppressWarnings("ALL")
 @RestController
 @RequestMapping("/api/item")
-public class TestController {
+public class DemoController {
     @Autowired
     private DemoService demoService;
 
-    public TestController(DemoService demoService){
+    public DemoController(DemoService demoService){
         this.demoService = demoService;
     }
     @GetMapping("/welcome")
@@ -52,8 +54,26 @@ public class TestController {
     @DeleteMapping("/del")
     public DebugHelp destroy(@RequestBody Map<String, List<Long>> body){
         List<Long> listOfId = body.get("itemId");
+        if (listOfId == null){return DebugHelp.fail("No Id Given");}
         DebugHelp debugHelp = new DebugHelp();
         debugHelp.addDebug("Test", demoService.deleteById(listOfId));
         return debugHelp;
     }
+
+    @PostMapping("/add")
+    public DebugHelp add(@Valid @RequestBody CreateItem createItem){
+        // still empty, still void
+         // add logic for throwing there
+        Item newItem = new Item(createItem);
+        demoService.insertItem(newItem);
+        return DebugHelp.fail("This is a Debug : " + newItem.getName() + " ," + newItem.getId() + " ," + newItem.getPrice() );
+    }
+
+    @PutMapping("/update")
+    public void update(@RequestBody Item item){
+        // still empty
+    }
 }
+
+
+// check the pasrsing system, see if
